get rid of useless variables to further simplify
I decided to get rid of the references to the items in the system
namespace. They weren't worth the small convenience. They would just
introduce bugs later on and complicate things imo. It also allowed me to
fuse the initialization of them into the registerSyncObjects function I
made last commit.