fix sync for root and implement RM command
The sync function is now part of the system namespace rather than a
member of the filesystem.
The filesystem contents will persist between sessions now.
Files and directories can now be deleted using the rm command.
The cat command now will output the contents of multiple files if they
are passed.
Added a reference to the parent path for file entries in the filesystem
(just to make things easier).