rewrote fileSystem.createFild and a couple fixes
fileSystem.createFile() has now been rewritten to be more generic.
It can create any number of files of any type with the proper properties
(as long as they're defined in the switch statement within).
It also uses the new typeTrans object to return human readable output.
Also implemented a small fix to the fileSystem.getPrefix() function in
the case where a lone filename was passed.